split pytorch example and copy code
