+++ examples/pytorch/README.md
@@ -1,0 +1,15 @@
+## Example Pytorch Model Training and Assessment
+
+### Contents
+
+* `dataset.py` - Creates synthetic test data.
+* `model.py` - Contains example Pytorch class.
+* `train_pytorch.py` - Trains example Pytorch model and saves it with Target class.
+* `attack_pytorch.py` - Loads the saved Pytorch model and runs attacks.
+
+### Usage
+
+```
+$ python -m train_pytorch
+$ python -m attack_pytorch
+```
